Updated README.md with instructions to run. Cleaned up some warning flags. Moved all code to an src folder.

+++ src/DynamicArray.h
@@ -18,7 +18,7 @@ public:
         data = nullptr;
     }
 
-    DynamicArray(int capacity) : capacity(capacity) {
+    DynamicArray(size_t capacity) : capacity(capacity) {
         data = new T[capacity];
     }
